refactor(tests): fold standalone test binaries into the integration harness
Each loose tests/*.rs file was its own crate and link step. Move all 21
into tests/integration/, merge tests/common's port helpers into the
harness common module, and point CI's web-gated step at the consolidated
binary.
