+++ tests/integration/agent_lifecycle_cli.rs
