+++ tests/integration/branch_exists_spawn_failure.rs
