+++ tests/integration/no_stale_doc_refs.rs
