+++ tests/integration/plugin_install.rs
