+++ tests/integration/serve_daemon_session_id_drain.rs
